Fix terminal
